password column added to users management
plain and hashed passwords are being stored in mongodb. plain passwords for purpose and hashed passwords for logics and functionalities. the plain password should not be shown or stored for security purposes, itis not recommended . and once a password is becrypted it cannot be decrypt due to becrypt security. it generates random hashes for passwords to make it more secure